traffic thread timers need to use uintmax_t vs int to allow for large values

+++ compat/signal.c
@@ -183,7 +183,7 @@ void disarm_itimer(void) {
 #endif
 }
 
-int set_itimer(int usecs) {
+int set_itimer(uintmax_t usecs) {
     int err = 0;
 #ifdef HAVE_SETITIMER
     if (usecs < 0) {
